refactor(storage): Extract CallbackManager to shared.rs and split leveldb.rs into directory module
Deduplicate identical callback storage/registration/execution code from all
three backends (memory, redb, leveldb) into a shared CallbackManager struct.
Split leveldb.rs (688 lines) into a directory module matching the memory/redb
pattern. Zero behavioral changes.
