+++ crates/storage/src/backends/mod.rs
@@ -43,9 +43,6 @@
 ///     let redb_store = RedbStore::open("/path/to/db")?;
 /// }
 /// ```
-// Memory backend uses tokio::sync::RwLock, only available on native platforms
-// For WASM, use the simplified memory store in the wasm crate
-#[cfg(not(target_arch = "wasm32"))]
 pub(crate) mod shared;
 
 #[cfg(not(target_arch = "wasm32"))]
@@ -70,11 +67,13 @@ pub mod opfs_env;
 #[cfg(all(test, not(target_arch = "wasm32")))]
 pub mod test_suite;
 
+pub use shared::CallbackCounts;
+
 #[cfg(not(target_arch = "wasm32"))]
 pub use memory::MemoryStore;
 
 #[cfg(all(feature = "redb", not(target_arch = "wasm32")))]
-pub use redb::{CallbackCounts, DurabilityMode, IntegrityReport, RedbStore, RedbStoreOptions};
+pub use redb::{DurabilityMode, IntegrityReport, RedbStore, RedbStoreOptions};
 
 #[cfg(all(target_arch = "wasm32", feature = "leveldb"))]
 pub use leveldb::LevelDbStore;
